buttons work!!!
completed the task of making the buttons for filtering the category of menu items work. Also styled them now I need to add CSS to the buttons depending on if the are active or not! Also I need to consolidate the Handler to change categories by passing value with the button clicked