fixed active menu filters
fixed the menu filters to show the active filter. Added more menu items to test the filter feature with more categories. categories are not scalable at this moment since the solution for showing active menu filters depends on changing the status for each button with one click. Need to find a way to use NavLink in a next.js app.