fixed footer
made footer stick to the bottom of the page and not be on top of everything else. Also made the top nav bar stick to the top an be on top of everything else. Also added a bottom border to the top bar and a top border to the footer.