MOBILE NAV ,finished media queries and mobile navigation
this is the best I can do for mobile nav. The only flaw in the system is that the website breaks when re-sizing the viewport life, because I cannot get a live viewport size, just the viewport size on mount. This is because the window object is not available in nextJS apps. When I find a way to do this I'll fix all of mobile vs desktop nav life conversion.