complete website
these are some possible upgrades to the site that I'll be working in the future

1. find a function that goes through an array and stops when it finds an object with that id then use it to determine the name of that item

2. find another function that counts how many elements with a certain id and use .legtht to output how many of them are in the car

MAYBE
last thing to do.
make buttons for each item have an id when that button is pushed it only affects the number of that item by itself, that way they are independent from eachother.

then the add item to cart button is pressed then push that object with id as a payload to the cartItems array